requirements.txt
# Core Django
Django==4.2.26
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Auth
djangorestframework-simplejwt==5.5.1
PyJWT==2.8.0

# Async & Background Jobs
celery==5.3.6
redis==5.0.1
flower==2.0.1

# Payments
stripe==8.2.0
razorpay==1.4.1

# Monitoring & Security
sentry-sdk==2.8.0
python-decouple==3.8

# Testing
pytest==8.0.1
pytest-django==4.8.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
ruff==0.2.2
black==24.3.0

# Security Scanning
safety==3.7.0
bandit==1.9.1
semgrep==1.143.1
radon==6.0.1
pip-audit==2.9.0
django-filter==24.3

# Utils
python-dateutil==2.8.2
pytz==2024.1
